requirements.txt
python-telegram-bot==13.7
pyyaml==6.0
redis==4.3.4
python-dateutil==2.8.2
requests==2.28.1